# src/main/resources/public/api-docs/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Mucry application API",
    "description" : "This is the murcy application API\n",
    "contact" : {
      "email" : "murcyapplication@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "/api"
  } ],
...
  "paths" : {
    "/user" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Create user",
        "description" : "Create user in the system",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "User logged created"
          },
          "400" : {
            "description" : "User already exists",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/confirm/{token}" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Confirm user register",
        "description" : "Confrim user register, from token send by email",
        "parameters" : [ {
          "name" : "token",
          "in" : "path",
          "description" : "token to be confirmed",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "User confirmed"
          },
          "404" : {
            "description" : "token not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/login" : {
      "post" : {
        "tags" : [ "user" ],
        "summary" : "Log in in the app",
        "description" : "Logs in in the application if the credentials are true",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "User logged in",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Auth"
                }
              }
            }
          },
          "401" : {
            "description" : "Invalid user, invalid password (or not confirmed user)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/user/info" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "Get current user info",
        "description" : "Get current logged user info",
        "responses" : {
          "201" : {
            "description" : "User data",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "user" ],
        "summary" : "Update current user info",
        "description" : "Update current user info (Only fill change fields)",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "User updated",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/user/info/{id}" : {
      "get" : {
        "tags" : [ "user" ],
        "summary" : "Get user info",
        "description" : "Get user info",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "user id to get info",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "User logged in",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "user" ],
        "summary" : "Update user info",
        "description" : "Update user info (Only fill change fields)",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "user id to update info",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "User logged in",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
    "/workflow/{id}/approve" : {
      "put" : {
        "tags" : [ "workflow" ],
        "summary" : "Aprrove workflow",
        "description" : "Approve workflow with given ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "workflow id to be approved",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WorkflowUpdateRequest"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Approved"
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
    "/workflow/{id}/deny" : {
      "put" : {
        "tags" : [ "workflow" ],
        "summary" : "Deny workflow",
        "description" : "Deny workflow with given ID",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "workflow id to be denied",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/WorkflowUpdateRequest"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "description" : "Denied"
          },
          "403" : {
            "description" : "Not authorized",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security" : [ {
          "bearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UserDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string",
            "description" : "user username"
          },
          "email" : {
            "type" : "string",
            "description" : "User email"
          },
          "fullName" : {
            "type" : "string",
            "description" : "User full name"
          },
          "role" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Rol"
            }
          }
        }
      },
...
      "Auth" : {
        "type" : "object",
        "properties" : {
          "jsonWebToken" : {
            "type" : "string"
          }
        }
      },
      "Error" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "integer"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "Rol" : {
        "type" : "string",
        "enum" : [ "USER", "EDITOR", "REVIEWER" ]
      },
...
      "LoginRequest" : {
        "required" : [ "password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "description" : "Username of the user to log in"
          },
          "password" : {
            "type" : "string",
            "description" : "Password of the user to log in"
          }
        }
      },
      "UpdateUserRequest" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "description" : "user username"
          },
          "password" : {
            "type" : "string",
            "description" : "User password"
          },
          "email" : {
            "type" : "string",
            "description" : "User email"
          },
          "fullName" : {
            "type" : "string",
            "description" : "User full name"
          },
          "role" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Rol"
            }
          }
        }
      },
...
      "UserRequest" : {
        "required" : [ "email", "fullName", "password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "description" : "user username"
          },
          "password" : {
            "type" : "string",
            "description" : "User password"
          },
          "email" : {
            "type" : "string",
            "description" : "User email"
          },
          "fullName" : {
            "type" : "string",
            "description" : "User full name"
          }
        }
      },
      "WorkflowUpdateRequest" : {
        "required" : [ "response" ],
        "type" : "object",
        "properties" : {
          "response" : {
            "type" : "string",
            "description" : "response message"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}